too long functions too many functions fix
